+++ pydivert/tests/test_windivert.py
@@ -13,107 +13,15 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-import itertools
-import socket
-import threading
 import time
 
 import pytest
 from pydivert.consts import Param
 from pydivert.windivert import WinDivert
 
-try:
-    # SocketServer has been renamed in python3 to socketserver
-    from socketserver import ThreadingMixIn, TCPServer, UDPServer, BaseRequestHandler
-    from queue import Queue
-except ImportError:
-    from SocketServer import ThreadingMixIn, TCPServer, UDPServer, BaseRequestHandler
-    from Queue import Queue
+from .fixtures import scenario, windivert_handle as w
 
-
-@pytest.fixture
-def w():
-    with WinDivert("false") as w:
-        yield w
-
-
-@pytest.fixture(params=list(itertools.product(
-    ("ipv4", "ipv6"),
-    ("tcp", "udp"),
-)), ids=lambda x: ",".join(x))
-def scenario(request):
-    ip_version, proto = request.param
-
-    """
-    if proto == socket.IPPROTO_TCP:
-        ServerBase = TCPServer
-        Handler = EchoUpperTCPHandler
-    else:
-        ServerBase = UDPServer
-        Handler = EchoUpperUDPHandler
-    class Server(ThreadingMixIn, ServerBase):
-        address_family = atype
-
-    server = Server((host, 0), Handler)
-    """
-
-    if ip_version == "ipv4":
-        atype = socket.AF_INET
-        host = "127.0.0.1"
-    else:
-        atype = socket.AF_INET6
-        host = "::1"
-    if proto == "tcp":
-        stype = socket.SOCK_STREAM
-    else:
-        stype = socket.SOCK_DGRAM
-
-    server = socket.socket(atype, stype)
-    server.bind((host, 0))
-    client = socket.socket(atype, stype)
-    client.bind((host, 0))
-
-    reply = Queue()
-
-    if proto == "tcp":
-        def server_echo():
-            server.listen(1)
-            conn, addr = server.accept()
-            conn.sendall(conn.recv(4096).upper())
-            conn.close()
-
-        def send(addr, data):
-            client.connect(addr)
-            client.sendall(data)
-            reply.put(client.recv(4096))
-    else:
-        def server_echo():
-            data, addr = server.recvfrom(4096)
-            server.sendto(data.upper(), addr)
-
-        def send(addr, data):
-            client.sendto(data, addr)
-            data, recv_addr = client.recvfrom(4096)
-            assert addr[:2] == recv_addr[:2]  # only accept responses from the same host
-            reply.put(data)
-
-    server_thread = threading.Thread(target=server_echo)
-    server_thread.start()
-
-    filt = "{proto}.SrcPort == {c_port} or {proto}.SrcPort == {s_port}".format(
-        proto=proto,
-        c_port=client.getsockname()[1],
-        s_port=server.getsockname()[1]
-    )
-
-    def send_thread(*args, **kwargs):
-        threading.Thread(target=send, args=args, kwargs=kwargs).start()
-        return reply
-
-    with WinDivert(filt) as w:
-        yield client.getsockname(), server.getsockname(), w, send_thread
-    client.close()
-    server.close()
+assert scenario, w  # keep fixtures
 
 
 def test_open():
